SGI:NONE add completed options
