+++ src/js/start.js
@@ -1,12 +1,26 @@
+import { keyDownEvent, clearCompletedEvent } from "./manageDom.js";
 export default class Start {
   constructor() {}
 
   init() {
     this.toDoList = document.querySelector(".todo-list");
     this.updateTask();
+    this.registerMainEvent();
   }
 
   updateTask() {
     this.toDoList.innerHTML = "";
+  }
+
+  registerMainEvent() {
+    const toDoInput = document.querySelector(".new-todo");
+    if (toDoInput) {
+      toDoInput.addEventListener("keydown", keyDownEvent);
+    }
+
+    const clearCompleted = document.querySelector(".clear-completed");
+    if (clearCompleted) {
+      clearCompleted.addEventListener("click", clearCompletedEvent);
+    }
   }
 }
